remove IsaAtomBase
